pyproject.toml
[tool.poetry]
name = "copro"
version = "1.10.0"
description = ""
authors = ["SOSP <sosp@halo.com>"]

[tool.poetry.dependencies]
python = "=3.8.10"
Django = "^4.0.2"
drf-yasg = { extras = ["validation"], version = "^1.20.0" }
configparser = "^5.2.0"
djangorestframework = "^3.13.1"
django-rest-swagger = "^2.2.0"
PyPDF2 = "^1.26.0"
Pillow = "^9.0.1"
opencv-python = "^4.5.5"
numpy = "^1.22.2"
pdf2image = "^1.16.0"
tqdm = "^4.63.0"
pytesseract = "^0.3.9"
datefinder = "^0.7.1"
google-api-core = "^2.5.0"
google-api-python-client = "^2.38.0"
google-cloud = "^0.34.0"
google-cloud-vision = "^2.7.0"
camelot-py = "^0.10.1"
PyMuPDF = "1.16.14"
pdfplumber = "^0.6.0"
tesseract = "^0.1.3"
tensorflow-addons = "^0.16.1"
python-doctr = "0.4.0"
imutils = "^0.5.4"
scikit-learn = "^1.1.1"
protobuf = "<3.20"
tensorflow-gpu = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
